Move moveCurUserEmailToFront and tests to trip-utils.
